requirements.txt
colorama==0.4.6
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.2
tomli==2.0.1
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
ipaddress==1.0.23
ipdata==3.4.6
kombu==5.3.5
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==23.2
passlib==1.7.4
peewee==3.17.1
pluggy==1.4.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic-core==2.16.3
pygments==2.17.2
PyMySQL==1.1.0
pytest==8.0.2
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.2
requests==2.31.0
rich==13.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomli==2.0.1
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
